Add SAME message decoding
Adds parsing to SAME events received over the serial port, including
location and event decoding.  This does wait for the entire event
including the end of message to be received.  In addition, if the
event fails to parse, it will try each successive message received
(since the transmission is sent three times) to parse the message.