Correct issues found on OS X
Mac OS X may return a null byte for the first character received;
to correct this, discard the null byte while detecting the EAR.
This doesn't seem to impact other read operations, however.

Also, OS X seems to ignore %j for anything but tm_yday with
strptime().  As a result, calculate the second offset in the year
from the date, and add that to the time.